Use 32 bits integers instead of 64

+++ apps/csharp/test_app/src/Controller.cs
@@ -25,12 +25,12 @@ public class Controller : ControllerBase
     }
 
     [HttpGet("getPrimesLessThan/{limit}")]
-    public IList<long> GetPrimesLessThan(long limit) {
+    public IList<int> GetPrimesLessThan(int limit) {
         return Primes.GetPrimesLessThan(limit);
     }
 
     [HttpGet("countPrimesLessThan/{limit}")]
-    public long CountPrimesLessThan(long limit) {
+    public int CountPrimesLessThan(int limit) {
         return Primes.CountPrimesLessThan(limit);
     }
 }
