+++ apps/csharp/test_app/src/Primes.cs
@@ -1,5 +1,5 @@
 internal static class Primes {
-    public static bool IsPrime(long num) {
+    public static bool IsPrime(int num) {
         if (num <= 1) return false;
         if (num == 2) return true;
         if (num % 2 == 0) return false; // the only even prime is 2
@@ -7,27 +7,27 @@ internal static class Primes {
         // Trial division (https://en.wikipedia.org/wiki/Prime_number#Trial_division)
         var limit = Math.Sqrt(num);
 
-        for (long i=3; i<=limit; i++) {
+        for (int i=3; i<=limit; i++) {
             if (num % i == 0) return false;
         }
 
         return true;
-    } 
+    }
 
-    public static IList<long> GetPrimesLessThan(long limit) {
-        var result = new List<long>();
+    public static IList<int> GetPrimesLessThan(int limit) {
+        var result = new List<int>();
 
-        for (long i=2; i<limit; i++) {
+        for (int i=2; i<limit; i++) {
             if (IsPrime(i)) result.Add(i);
         }
 
         return result;
     }
 
-    public static long CountPrimesLessThan(long limit) {
-        long result = 0;
+    public static int CountPrimesLessThan(int limit) {
+        int result = 0;
 
-        for (long i=2; i<limit; i++) {
+        for (int i=2; i<limit; i++) {
             if (IsPrime(i)) result++;
         }
